Split up events and their related methods
