+++ src/main/kotlin/com/kuvaszuptime/kuvasz/models/ServiceError.kt
