+++ src/main/kotlin/com/kuvaszuptime/kuvasz/repositories/UptimeEventRepository.kt
@@ -3,10 +3,10 @@ package com.kuvaszuptime.kuvasz.repositories
 import arrow.core.toOption
 import com.kuvaszuptime.kuvasz.models.MonitorDownEvent
 import com.kuvaszuptime.kuvasz.models.UptimeMonitorEvent
-import com.kuvaszuptime.kuvasz.models.toUptimeStatus
 import com.kuvaszuptime.kuvasz.tables.UptimeEvent.UPTIME_EVENT
 import com.kuvaszuptime.kuvasz.tables.daos.UptimeEventDao
 import com.kuvaszuptime.kuvasz.tables.pojos.UptimeEventPojo
+import com.kuvaszuptime.kuvasz.util.toUptimeStatus
 import org.jooq.Configuration
 import java.time.OffsetDateTime
 import javax.inject.Inject
