+++ src/test/kotlin/com/kuvaszuptime/kuvasz/events/EventTest.kt
@@ -3,9 +3,9 @@ package com.kuvaszuptime.kuvasz.events
 import arrow.core.Option
 import com.kuvaszuptime.kuvasz.enums.UptimeStatus
 import com.kuvaszuptime.kuvasz.models.MonitorUpEvent
-import com.kuvaszuptime.kuvasz.models.getEndedEventDuration
-import com.kuvaszuptime.kuvasz.models.runWhenStateChanges
 import com.kuvaszuptime.kuvasz.tables.pojos.UptimeEventPojo
+import com.kuvaszuptime.kuvasz.util.getEndedEventDuration
+import com.kuvaszuptime.kuvasz.util.runWhenStateChanges
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
 import io.micronaut.http.HttpStatus
